Rename package noos-tf to publish to pypi
